Changed the seed to be uint64_t, it seems from testing knuth_b really is decent maybe

+++ parameters.h
@@ -190,7 +190,7 @@ struct Parameters {
    static bool refineOnRestart;
    static bool forceRefinement;
    static bool shouldFilter;
-   static int seed;
+   static uint64_t seed;
    static bool useAlpha1;
    static Real alpha1RefineThreshold;
    static Real alpha1CoarsenThreshold;
